Refactor ScannBlacklistCommand and handler for multiple scanners
Simplify ScannBlacklistCommand initialization using primary constructor syntax. Update handler to support multiple IBlacklistScanner instances, iterating through each to scan domains. Optimize dependencies by removing unused fields and streamlining logic.
