--- src/DNS-BLM.Application/Services/IBlacklistScanner.cs
@@ -1,6 +1,0 @@
-namespace DNS_BLM.Application.Services;
-
-public interface IBlacklistScanner
-{
-    public Task<List<string>> Scan(List<string> domains);
-}
