+++ src/DNS-BLM.Infrastructure/Services/ServiceInterfaces/IBlacklistScanner.cs
@@ -1,0 +1,9 @@
+using DNS_BLM.Infrastructure.Dtos;
+
+namespace DNS_BLM.Infrastructure.Services.ServiceInterfaces;
+
+public interface IBlacklistScanner
+{
+    public Task Scan(List<string> domains);
+    public string ScannerName { get; }
+}
